+++ src/dsp/afsk1200/cafsk12.cpp
@@ -54,12 +54,12 @@ void CAfsk12::reset()
     for (f = 0, i = 0; i < CORRLEN; i++) {
         corr_mark_i[i] = cos(f);
         corr_mark_q[i] = sin(f);
-        f += 2.0*M_PI*FREQ_MARK/FREQ_SAMP;
+        f += 2*(float)M_PI*FREQ_MARK/FREQ_SAMP;
     }
     for (f = 0, i = 0; i < CORRLEN; i++) {
         corr_space_i[i] = cos(f);
         corr_space_q[i] = sin(f);
-        f += 2.0*M_PI*FREQ_SPACE/FREQ_SAMP;
+        f += 2*(float)M_PI*FREQ_SPACE/FREQ_SAMP;
     }
 
 }
@@ -460,4 +460,3 @@ void CAfsk12::ax25_disp_packet(unsigned char *bp, unsigned int len)
         emit newMessage(message);
     }
 }
-
